// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as auth from "../auth.js";
import type * as http from "../http.js";
import type * as questionPapers from "../questionPapers.js";
import type * as questions from "../questions.js";
import type * as questionsInternal from "../questionsInternal.js";
import type * as router from "../router.js";
import type * as sections from "../sections.js";
import type * as syllabusFiles from "../syllabusFiles.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  http: typeof http;
  questionPapers: typeof questionPapers;
  questions: typeof questions;
  questionsInternal: typeof questionsInternal;
  router: typeof router;
  sections: typeof sections;
  syllabusFiles: typeof syllabusFiles;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;